Added assertNotThrows function
The `assertNotThrows` function is generating an error whenever the autoclosure passed to it is throwing an error.

This function is mainly written to test that a function is not throwing an error, without testing the returned value if there is one.
For the other cases, the `assertThat` function should be preferred.